updates the `kubectl_manifest` to use a `file` resource instead of `templatefile`
the `infra/account.hcl` file now includes an `aws_caller_arn` local to get the aws caller identity arn. this ensures that the resource is properly configured and the identity of the caller is correctly identified.

the `kubectl_manifest` resource in `infra/modules/kubevela-addons/definitions.tf` now uses a `file` resource instead of a `templatefile` resource. this ensures that the resource is properly configured and the file is properly read.